+++ src/router/v1/recommend/helper.ts
@@ -1,5 +1,5 @@
 import { getImgUrl } from "../../../utils/img.js";
-import { generateRecipeKey } from "../recipes/helper.js";
+import { generateRecipeKey } from "../../../services/recipes/helper.js";
 import { HomeRecommendRecipe, RecommendRecipe } from "./type.js";
 
 const formatRecipeData = (recipe: RecommendRecipe) => {
