+++ src/services/recipes/recipe/updateRecipeService.ts
@@ -6,7 +6,7 @@ import { EditRecipe } from "./type.js";
 import { RecipeInfo } from "./type.js";
 import { ServiceError } from "../../helpers/ServiceError.js";
 import { lightSlugify } from "../../../utils/normalize.js";
-import { getNewProductData } from "../../../router/v1/recipes/helper.js";
+import { getNewProductData } from "../helper.js";
 import { IngredientNewProduct } from "../../../router/v1/recipes/recipes.js";
 import { arrayToPlaceholders } from "../../../utils/query.js";
 import { UpdateRecipeServiceParams } from "./type.js";
