+++ src/services/recipes/recipesCreateService/helper.ts
@@ -4,10 +4,7 @@ import {
   INewRecipe,
   IngredientNewProduct,
 } from "../../../router/v1/recipes/recipes.js";
-import {
-  getNewProductData,
-  getNewRecipeData,
-} from "../../../router/v1/recipes/helper.js";
+import { getNewProductData, getNewRecipeData } from "../helper.js";
 import { lightSlugify } from "../../../utils/normalize.js";
 
 export const insertRecipe = async (
