+++ src/services/recipes/recipesCreateService/index.ts
@@ -5,7 +5,7 @@ import mysqlDB from "../../../db/mysql.js";
 import { SerializedUser } from "../../../config/passport.js";
 
 import { INewRecipe } from "../../../router/v1/recipes/recipes.js";
-import { generateRecipeKey } from "../../../router/v1/recipes/helper.js";
+import { generateRecipeKey } from "../helper.js";
 import { CreateRecipeBody } from "../../../controllers/recipes/recipesCreateController.js";
 import {
   insertIngredients,
